% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4ra-package.R
\docType{package}
\name{m4ra-package}
\alias{m4ra-package}
\alias{_PACKAGE}
\title{m4ra: Many-to-Many Multi-Modal Routing Aggregator}
\description{
Many-to-many multi-modal routing aggregator, combining 'dodgr' and 'gtfsrouter' for complete routing through urban networks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UrbanAnalyst/m4ra}
  \item Report bugs at \url{https://github.com/UrbanAnalyst/m4ra/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
