% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4ra-function.R
\name{m4ra}
\alias{m4ra}
\title{Many-to-many multi-modal routing function}
\usage{
m4ra(net, gtfs = NULL, from = NULL, to = NULL, quiet = FALSE)
}
\arguments{
\item{net}{An urban network in \code{osmdata_sc} format from the \pkg{osmdata}
package.}

\item{gtfs}{An optional GTFS feed in \code{gtfs} format from the \pkg{gtfsrouter}
package.}

\item{from}{A vector of start points for routing, specified as Open Street
Map (OSM) identifier ("ID") values, generally from the ID column of
\code{dodgr_vertices(net)}, or obtained by matching coordinates to the network
with the \pkg{dodgr} \code{match_pts_to_verts} function.}

\item{to}{An optional vector of end points for routing.}

\item{quiet}{If \code{FALSE}, display progress information on screen.}
}
\value{
(Unsure at this stage)
}
\description{
Perform a many-to-many multi-modal routing query
}
\seealso{
Other main: 
\code{\link{m4ra_gtfs_traveltimes}()},
\code{\link{m4ra_times}()}
}
\concept{main}
