% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight-networks.R
\name{m4ra_batch_weight_networks}
\alias{m4ra_batch_weight_networks}
\title{Cache a directory full of street networks for routing queries}
\usage{
m4ra_batch_weight_networks(net_dir, remove_these = NULL)
}
\arguments{
\item{net_dir}{Path of local directory containing 'sc'-format street
networks.}

\item{remove_these}{Names of any 'sc'-format files which should not be
converted into weighted network form.}
}
\value{
A character vector of local locations of cached versions of the
variously weighted network representations used in \link{m4ra}.
}
\description{
This function runs of a directory which contain a number of \pkg{silicate} or
\code{sc}-formatted street networks, generated with the \code{dodgr_streetnet_sc}
function of the \pkg{dodgr} package. The function uses a default cache
location specified by \code{rappdirs::user_cache_dir()}. This location can be
over-ridden by specifying a local environment variable, "M4RA_CACHE_DIR".
}
\seealso{
Other cache: 
\code{\link{m4ra_network_hash}()},
\code{\link{m4ra_weight_networks}()}
}
\concept{cache}
