% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times-bike-car.R
\name{m4ra_bike_car_ratio_areas}
\alias{m4ra_bike_car_ratio_areas}
\title{Convert bicycle and automobile times into equivalent areas for a specified
range of ratios of the two travel times.}
\usage{
m4ra_bike_car_ratio_areas(bike_car_dat, ratio_lims = 1:20/5)
}
\arguments{
\item{bike_car_dat}{Result of \link{m4ra_bike_car_times} function.}

\item{ratio_lims}{Vector of ratio limits used to calculate areas within which
ratio of bicycle to automobile times is less than specified limit.}
}
\description{
Convert bicycle and automobile times into equivalent areas for a specified
range of ratios of the two travel times.
}
\seealso{
Other analyses: 
\code{\link{m4ra_bike_car_times}()}
}
\concept{analyses}
