% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times-bike-car.R
\name{m4ra_bike_car_times}
\alias{m4ra_bike_car_times}
\title{Calculate times from a specified point by bicycle and automobile, along with
corresponding walking distances.}
\usage{
m4ra_bike_car_times(city = NULL, from = NULL)
}
\arguments{
\item{city}{City for which values are to be calculated}

\item{from}{Vector of OSM ID values of vertices from which ratios are to be
calculated. Typically obtained by loading one weighted network, and
sampling or extracting vertices from the \pkg{dodgr} function
\code{dodgr_vertices}.}

\item{ratio_limits}{Limits of ratios of bicycle to automobile times for which
areas are to be calculated.}
}
\value{
A \code{data.frame} of destination vertices, including Open Street Map ID
values, and longitude and latitude values, and four variables:
\itemize{
\item 'car_t' Times in seconds to travel with automobile.
\item 'bike_t' Times in seconds to travel by bicycles.
\item 'walk_d' Equivalent walking distance in kilometres.
\item 'ratio' Ratio of bicycle to automobile travel times.
}
}
\description{
This city expects weighted networks to have been generated with the
\link{m4ra_batch_weight_networks} function, and for corresponding networks to
exist in the \code{m4ra} cache directory for the specified city.
}
\examples{
\dontrun{
city <- "<city_name>"
net <- m4ra_load_cached_network (city = city, mode = "foot")
v <- dodgr::dodgr_vertices (net)
from <- sample (v$id, size = 10)
dat <- m4ra_bike_car_ratios (city = city, from = from)
}
}
\concept{analyses}
