% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{m4ra_cache_network}
\alias{m4ra_cache_network}
\title{Save a weighted street network in both full and contracted forms to local
cache.}
\usage{
m4ra_cache_network(graph, city, mode)
}
\arguments{
\item{graph}{The full network to be cached.}

\item{city}{Name of city; used to name cached network files.}

\item{mode}{Mode of transport, as name of weighting profile used to weight
networks in \pkg{dodgr}; used to name cached network files.}
}
\description{
This uses \pkg{fst}, which strips all attributes, so they are saved
separately, and re-attached in \code{load_cached_network()}.
}
\seealso{
Other cache: 
\code{\link{m4ra_batch_weight_networks}()},
\code{\link{m4ra_load_cached_network}()},
\code{\link{m4ra_network_hash}()},
\code{\link{m4ra_weight_networks}()}
}
\concept{cache}
