% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists-n-pts.R
\name{m4ra_dists_n_pts}
\alias{m4ra_dists_n_pts}
\title{Calculate network distances to nearest 'n' points.}
\usage{
m4ra_dists_n_pts(graph, from = NULL, to = NULL, npts = 3L)
}
\arguments{
\item{graph}{A \code{dodgr} network returned from the \link{weight_streetnet}
function using a network obtained with the \pkg{osmdata} \code{osmdata_sc}
function.}

\item{from}{Vector of OpenStreetMap ID values from which route distances are
to be calculated.}

\item{to}{Vector of OpenStreetMap ID values to which nearest 'npts' distances are
to be calculated.}

\item{npts}{Number of points to which distances are to be calculated.}
}
\description{
Calculate network distances to nearest 'n' points.
}
\seealso{
Other main: 
\code{\link{m4ra_gtfs_traveltimes}()},
\code{\link{m4ra_times_multi_mode}()},
\code{\link{m4ra_times_single_mode}()}
}
\concept{main}
