% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{m4ra_gtfs_traveltimes}
\alias{m4ra_gtfs_traveltimes}
\title{Construct a travel time matrix between all pairs of stops in a 'GTFS' feed.}
\usage{
m4ra_gtfs_traveltimes(gtfs, start_time_limits, day)
}
\arguments{
\item{gtfs}{A 'GTFS' feed extracted with the \pkg{gtfsrouter} function,
'extract_gtfs'.}

\item{start_time_limits}{As for the 'gtfs_traveltimes' function of
\pkg{gtfsrouter}, a vector of two values specifying the earliest and latest
departure times from each station.}

\item{day}{As for the 'gtfs_traveltimes' function of \pkg{gtfsrouter}, the
day for which the matrix is to be calculated.}
}
\value{
A list of two integer matrices:
\itemize{
\item The fastest travel times between all pairs of stops for the specified
'start_time_limits'; and
\item The corresponding numbers of transfers.
}
}
\description{
Construct a travel time matrix between all pairs of stops in a 'GTFS' feed.
}
\seealso{
Other main: 
\code{\link{m4ra_dists_n_pts}()},
\code{\link{m4ra_times_multi_mode}()},
\code{\link{m4ra_times_single_mode}()}
}
\concept{main}
