% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4ra-package.R
\docType{data}
\name{m4ra_hampi}
\alias{m4ra_hampi}
\title{m4ra_hampi}
\format{
A \pkg{silicate} \code{sc} representation of the street
network of Hampi.
}
\description{
A sample street network from the township of Hampi, Karnataka, India. The
same as the \code{dodgr::hampi} dataset, but in \code{sc} rather than \code{sf} format.
}
\note{
Can be re-created with the following commands.
}
\examples{
\dontrun{
hampi <- dodgr_streetnet_sc ("hampi india")
}
}
\concept{data}
\keyword{datasets}
