% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{m4ra_load_cached_network}
\alias{m4ra_load_cached_network}
\title{Load cached file for one city and mode}
\usage{
m4ra_load_cached_network(
  city = NULL,
  mode = "foot",
  contracted = TRUE,
  filename = NULL
)
}
\arguments{
\item{city}{City for which file is to be loaded.}

\item{mode}{One of "foot", "bicycle", or "motorcar".}

\item{contracted}{If \code{TRUE}, load the contracted version of the graph;
otherwise load the full graph.}

\item{filename}{Optional name of specific file to load.}
}
\value{
Previously cached, weighted streetnet for specified city and mode.
}
\description{
Load cached file for one city and mode
}
\seealso{
Other cache: 
\code{\link{m4ra_batch_weight_networks}()},
\code{\link{m4ra_cache_network}()},
\code{\link{m4ra_network_hash}()},
\code{\link{m4ra_weight_networks}()}
}
\concept{cache}
