% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times-multi-modal.R
\name{m4ra_times_mm_car}
\alias{m4ra_times_mm_car}
\title{Calculate relative times from a specified point by multi-modal transport
versus private automobile, along with (optionally) corresponding walking
distances.}
\usage{
m4ra_times_mm_car(
  net_sc = NULL,
  gtfs = NULL,
  city_name = NULL,
  day = NULL,
  start_time_limits = NULL,
  initial_mode = "foot",
  final_mode = "foot",
  from = NULL,
  fast = FALSE,
  n_closest = 10L,
  walk_dists = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{net_sc}{Local file path to a \pkg{silicate}, "SC", format object
containing network data used to generate weighted street networks.}

\item{gtfs}{Local file path to a \pkg{gtfsrouter} object saved via \code{saveRDS},
containing GTFS (General Transit Feed Specification) data for network
represented in \code{net_sc}. This \code{.Rds} object may include additional timetable
or transfer information in addition to data represented in original
\code{zip}-format data of provided GTFS feed.}

\item{city_name}{Name of city used to name cached files.}

\item{day}{As for the 'gtfs_traveltimes' function of \pkg{gtfsrouter}, the
day for which the matrix is to be calculated.}

\item{start_time_limits}{As for the 'gtfs_traveltimes' function of
\pkg{gtfsrouter}, a vector of two values specifying the earliest and latest
departure times from each station.}

\item{initial_mode}{Initial mode of transport from origin points towards
public transport stop.}

\item{final_mode}{The mode of transport used for the final stage from GTFS
stops to destination points.}

\item{from}{Vector of OSM ID values of vertices from which ratios are to be
calculated. Typically obtained by loading one weighted network, and
sampling or extracting vertices from the \pkg{dodgr} function
\code{dodgr_vertices}.}

\item{fast}{Values of \code{TRUE} generate potentially enormous matrices which may
not fit in local memory. The default value of \code{FALSE} is generally safer, but
calculation may take considerably longer.}

\item{n_closest}{Final travel times to each destination point are calculated
by tracing back times to this number of closest GTFS stops. Lower values will
result in faster calculation times, yet with potentially inaccurate results.}

\item{walk_dists}{If \code{TRUE}, also calculate equivalent walking distances.}

\item{quiet}{If \code{FALSE}, display progress information on screen.}
}
\value{
A \code{data.frame} of destination vertices, including Open Street Map ID
values, and longitude and latitude values, and four variables:
\itemize{
\item 'car_t' Times in seconds to travel with automobile.
\item 'mm_t' Times in seconds for equivalent multi-modal transport.
\item 'walk_d' Equivalent walking distance in kilometres.
\item 'ratio' Ratio of multi-modal to automobile travel times.
}
}
\description{
This city expects weighted networks to have been generated with the
\link{m4ra_batch_weight_networks} function, and for corresponding networks to
exist in the \code{m4ra} cache directory for the specified city.
}
\examples{
\dontrun{
city <- "<city_name>"
net <- m4ra_load_cached_network (city = city, mode = "foot")
v <- dodgr::dodgr_vertices (net)
from <- sample (v$id, size = 10)
dat <- m4ra_mm_car_ratios (city = city, from = from)
}
}
\seealso{
Other analyses: 
\code{\link{m4ra_bike_car_ratio_areas}()},
\code{\link{m4ra_times_bike_car}()}
}
\concept{analyses}
