% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times-multi-modal.R
\name{m4ra_times_multi_modal}
\alias{m4ra_times_multi_modal}
\title{Calculate matrix of pair-wise travel times between points using multiple
modes of transport.}
\usage{
m4ra_times_multi_modal(graph, gtfs, from = NULL, start_time_limits = NULL)
}
\arguments{
\item{graph}{A \code{dodgr} network returned from the \link{weight_streetnet}
function using a network obtained with the \pkg{osmdata} \code{osmdata_sc}
function.}

\item{gtfs}{A 'GTFS' feed extracted with the \pkg{gtfsrouter} function,
'extract_gtfs'.}

\item{from}{Vector or matrix of points \strong{from} which route distances are to
be calculated. If not given, times are calculated from all points in the
network.}

\item{start_time_limits}{As for the 'gtfs_traveltimes' function of
\pkg{gtfsrouter}, a vector of two values specifying the earliest and latest
departure times from each station.}
}
\description{
Alias for \link{m4ra_times_multi_mode}
}
\seealso{
Other main: 
\code{\link{m4ra_gtfs_traveltimes}()},
\code{\link{m4ra_prepare_data}()},
\code{\link{m4ra_times_multi_mode}()},
\code{\link{m4ra_times_single_mode}()},
\code{\link{m4ra_times_to_gtfs_stops}()},
\code{\link{m4ra}()}
}
\concept{main}
