% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times-single-mode.R
\name{m4ra_times_single_mode}
\alias{m4ra_times_single_mode}
\title{Calculate matrix of pair-wise travel times between points using a single mode
of transport.}
\usage{
m4ra_times_single_mode(graph, from = NULL, to = NULL, path = NULL)
}
\arguments{
\item{graph}{A \code{dodgr} network returned from the \link{weight_streetnet}
function using a network obtained with the \pkg{osmdata} \code{osmdata_sc}
function.}

\item{from}{Vector or matrix of points \strong{from} which route distances are to
be calculated (see Notes)}

\item{to}{Vector or matrix of points \strong{to} which route distances are to be
calculated (see Notes)}

\item{path}{If specified, save individual travel time vectors for each 'from'
point at that local directory.

\code{from} and \code{to} values can be either two-column matrices or
equivalent of longitude and latitude coordinates, or else single columns
precisely matching node numbers or names given in \code{graph$from} or
\code{graph$to}. If \code{to} is \code{NULL}, pairwise distances are calculated
between all points specified in \code{from}. If both \code{from} and \code{to} are \code{NULL},
pairwise distances are calculated between all nodes in \code{graph}.}
}
\description{
Calculate matrix of pair-wise travel times between points using a single mode
of transport.
}
\examples{
net <- dodgr::weight_streetnet (m4ra_hampi, wt_profile = "foot")
traveltimes <- m4ra_times_single_mode (net)
}
\seealso{
Other main: 
\code{\link{m4ra_gtfs_traveltimes}()},
\code{\link{m4ra_prepare_data}()},
\code{\link{m4ra_times_multi_modal}()},
\code{\link{m4ra_times_multi_mode}()},
\code{\link{m4ra_times_to_gtfs_stops}()}
}
\concept{main}
