% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{m4ra_times_to_gtfs_stops}
\alias{m4ra_times_to_gtfs_stops}
\title{Construct a travel time matrix to or from all stops in a 'GTFS' feed from or
to to all points in a street network.}
\usage{
m4ra_times_to_gtfs_stops(graph, gtfs, graph_to_gtfs = TRUE)
}
\arguments{
\item{graph}{A \code{dodgr} network returned from the \link{weight_streetnet}
function using a network obtained with the \pkg{osmdata} \code{osmdata_sc}
function.}

\item{gtfs}{A 'GTFS' feed extracted with the \pkg{gtfsrouter} function,
'extract_gtfs'.}

\item{graph_to_gtfs}{If \code{TRUE}, generate matrix of times from all network
junctions in 'graph' to each stop in the 'gtfs$stops' table; otherwise
generate matrix of times from all stops to all network junctions.}
}
\value{
An integer matrix of fastest travel times either between all 'gtfs'
stops and all network points (for 'graph_to_gtfs = FALSE'), or the other way
around (for 'graph_to_gtfs = TRUE').
}
\description{
Construct a travel time matrix to or from all stops in a 'GTFS' feed from or
to to all points in a street network.
}
\seealso{
Other main: 
\code{\link{m4ra_gtfs_traveltimes}()},
\code{\link{m4ra_times_single_mode}()},
\code{\link{m4ra}()}
}
\concept{main}
