% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertices.R
\name{m4ra_vertices}
\alias{m4ra_vertices}
\title{m4ra_vertices}
\usage{
m4ra_vertices(graph, city)
}
\arguments{
\item{graph}{A flat table of graph edges. Must contain columns labelled
\code{from} and \code{to}, or \code{start} and \code{stop}. May also contain
similarly labelled columns of spatial coordinates (for example
\code{from_x}) or \code{stop_lon}).}

\item{city}{Name of city; used to name cached network files.}
}
\value{
A \code{data.frame} of vertices with unique numbers (\code{n}).
}
\description{
Extract vertices of graph, including spatial coordinates if included.
Modified version of equivalent \pkg{dodgr} function that uses different hash
comparisons, and uses \pkg{fst} for write/read functions. This version also
does none of the checks implemented in the \pkg{dodgr} version, and assumes
graphs remain unmodified throughout.
}
\note{
Values of \code{n} are 0-indexed
}
\examples{
graph <- dodgr::weight_streetnet (dodgr::hampi)
v <- m4ra_vertices (graph, "hampi")
}
\concept{misc}
