% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight-networks.R
\name{m4ra_weight_networks}
\alias{m4ra_weight_networks}
\title{Cache weighted networks for routing query}
\usage{
m4ra_weight_networks(net, quiet = TRUE)
}
\arguments{
\item{net}{An urban network in \code{osmdata_sc} format from the \pkg{osmdata}
package.}

\item{quiet}{If \code{FALSE}, display progress information on screen.}
}
\value{
A character vector of local locations of cached versions of the
variously weighted network representations used in \link{m4ra}.
}
\description{
Uses a default cache location specified by \code{rappdirs::user_cache_dir()}. This
location can be over-ridden by specifying a local environment variable,
"M4RA_CACHE_DIR".
}
\seealso{
Other cache: 
\code{\link{m4ra_network_hash}()}
}
\concept{cache}
