% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4ra-package.R
\docType{package}
\name{m4ra-package}
\alias{m4ra-package}
\alias{_PACKAGE}
\title{m4ra: Many-to-Many Multi-Modal Routing Engine}
\description{
Many-to-~any ~ulti~Modal ~outing ~ngine.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ATFutures/m4ra}
  \item Report bugs at \url{https://github.com/ATFutures/m4ra/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
