% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{m4ra_gtfs_traveltimes}
\alias{m4ra_gtfs_traveltimes}
\title{Construct a travel time matrix between all pairs of stops in a 'GTFS' feed.}
\usage{
m4ra_gtfs_traveltimes(gtfs, start_time_limits, day, next_interval = FALSE)
}
\arguments{
\item{gtfs}{A 'GTFS' feed extracted with the \pkg{gtfsrouter} function,
'extract_gtfs'.}

\item{start_time_limits}{As for the 'gtfs_traveltimes' function of
\pkg{gtfsrouter}, a vector of two values specifying the earliest and latest
departure times from each station.}

\item{day}{As for the 'gtfs_traveltimes' function of \pkg{gtfsrouter}, the
day for which the matrix is to be calculated.}

\item{next_interval}{If \code{TRUE}, calculate time intervals to subsequent trips
after identified fastest trips. These subsequent trips may not necessarily be
as fast as the initial trips, and so this requires a second calculation of
all travel times. Setting this parameter to \code{TRUE} will therefore generally
double the calculation time.}
}
\value{
A list of two or three integer matrices:
\itemize{
\item "duration": The fastest travel times between all pairs of stops for the specified
'start_time_limits'; and
\item "ntransfers": The corresponding numbers of transfers.
\item (Only if 'next_interval = TRUE') "intervals": a matrix of intervals (in
seconds) until the next fastest service after that corresponding to the times
in the 'duration' item.
}
}
\description{
Construct a travel time matrix between all pairs of stops in a 'GTFS' feed.
}
\seealso{
Other main: 
\code{\link{m4ra_dists_n_pts}()},
\code{\link{m4ra_times_multi_mode}()},
\code{\link{m4ra_times_single_mode}()}
}
\concept{main}
