% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{m4ra_network_hash}
\alias{m4ra_network_hash}
\title{Return a unique hash for the input network.}
\usage{
m4ra_network_hash(graph)
}
\arguments{
\item{graph}{A \pkg{silicate}, "SC", format object containing network data
used to generate weighted street networks.}
}
\value{
Single character value with unique hash of given network.
}
\description{
This can be used to identify cached versions of a given network, in
particular to determine whether or not weighted versions of a given network
have been previously cached.
}
\seealso{
Other cache: 
\code{\link{m4ra_batch_weight_networks}()},
\code{\link{m4ra_cache_network}()},
\code{\link{m4ra_load_cached_network}()},
\code{\link{m4ra_weight_networks}()}
}
\concept{cache}
