% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parking.R
\name{m4ra_parking}
\alias{m4ra_parking}
\title{Get data on parking availability to estimate time penalties for automobile
travel.}
\usage{
m4ra_parking(bb, city_name, mode = "foot")
}
\arguments{
\item{bb}{Bounding box of city for query to extract parking data.}

\item{city_name}{Name of city used to name cached files.}

\item{mode}{Mode of transport used to extract OSM node IDs at which to
estimate relative parking availability.}
}
\value{
A \code{data.frame} of the vertices of the (contracted) network, with
additional columns quantifying number of parking spaces associated with each
vertex, as well as the total volume of all surrounding buildings.
}
\description{
Get data on parking availability to estimate time penalties for automobile
travel.
}
