% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parking.R
\name{m4ra_parking}
\alias{m4ra_parking}
\title{Get data on parking availability and building volumes to estimate time
penalties for automobile travel.}
\usage{
m4ra_parking(bb, city_name, mode = "foot", dlim = 5000, k = 1000)
}
\arguments{
\item{bb}{Bounding box of city for query to extract parking data.}

\item{city_name}{Name of city used to name cached files.}

\item{mode}{Mode of transport used to extract OSM node IDs at which to
estimate relative parking availability.}

\item{dlim}{Distance limit in metres out to which contributions of parking
and buildings will be aggregated.}

\item{k}{With of exponential function used to weight contributions with
distance, \code{exp(-d / k)} for distance \code{d}. Default value of 1000 metres
decreases weight to 37\% at 1km, 14\% at 2km, and 0.6\% at 5km.}
}
\value{
A \code{data.frame} of the vertices of the (contracted) network, with
additional columns quantifying number of parking spaces associated with each
vertex, as well as the total volume of all surrounding buildings.
}
\description{
Get data on parking availability and building volumes to estimate time
penalties for automobile travel.
}
\seealso{
Other analyses: 
\code{\link{m4ra_bike_car_ratio_areas}()},
\code{\link{m4ra_bike_car_times}()}
}
\concept{analyses}
