% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{m4ra_prepare_data}
\alias{m4ra_prepare_data}
\title{Prepare network and GTFS components for 'm4ra' queries.}
\usage{
m4ra_prepare_data(
  net_sc = NULL,
  gtfs = NULL,
  start_time_limits = NULL,
  day = NULL,
  city_name = NULL
)
}
\arguments{
\item{net_sc}{Local file path to a \pkg{silicate}, "SC", format object
containing network data used to generate weighted street networks.}

\item{gtfs}{Local file path to a \pkg{gtfsrouter} object saved via \code{saveRDS},
containing GTFS (General Transit Feed Specification) data for network
represented in \code{net_sc}. This \code{.Rds} object may include additional timetable
or transfer information in addition to data represented in original
\code{zip}-format data of provided GTFS feed.}

\item{start_time_limits}{As for the 'gtfs_traveltimes' function of
\pkg{gtfsrouter}, a vector of two values specifying the earliest and latest
departure times from each station.}

\item{day}{As for the 'gtfs_traveltimes' function of \pkg{gtfsrouter}, the
day for which the matrix is to be calculated.}

\item{city_name}{Name of city used to name cached files.}
}
\description{
This is a meta-function that calls the following functions:
\itemize{
\item \link{m4ra_weight_networks} to generate several versions of the input
street network weighted for different kinds of transport.
\item \link{m4ra_gtfs_traveltimes} to generate
}
This stores several differently-weighted version of the street networks in a
local cache directory defined by the \code{user_cache_dir()} function of the
\pkg{rappdirs} package, or by a local environment parameter,
"M4RA_CACHE_DIR", if set.
}
